# specs/002-web-app/contracts/todo-web-api-contract.yaml
openapi: 3.0.0
info:
  title: Todo Web Application API
  description: API contract for the Todo Web Application (Phase II)
  version: 1.0.0
servers:
  - url: https://api.todo-web-app.com/v1
    description: Production server
  - url: https://staging-api.todo-web-app.com/v1
    description: Staging server
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  description: User's password
                  example: "securePassword123"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    description: Unique identifier for the user
                    example: "550e8400-e29b-41d4-a716-446655440000"
                  email:
                    type: string
                    format: email
                    description: User's email address
                    example: "user@example.com"
                  created_at:
                    type: string
                    format: date-time
                    description: Account creation timestamp
                    example: "2023-01-01T00:00:00Z"
        '400':
          description: Invalid input or user already exists
  /auth/login:
    post:
      summary: Login a user
      description: Authenticate a user and return session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  description: User's password
                  example: "securePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  token_type:
                    type: string
                    description: Type of token
                    example: "bearer"
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2023-01-01T00:00:00Z"
...